test: custom receiver in bridge order
